notes - refactoring code
