+++ src/services/localStorage.ts
@@ -1,5 +1,3 @@
-import {getNotes} from '../widgets/notes/service/getNotes';
-
 export const setDataToLocalStorage = (localStorageName: string, stringData: string) => {
     return localStorage.setItem(localStorageName, stringData);
 };
