+++ src/widgets/notes/components/accordionItem/AccordionItem.module.css
