+++ src/widgets/notes/components/accordionItem/typings/index.ts
@@ -1,0 +1,9 @@
+import { NotesType } from '../../../typings';
+
+export type PropsType = {
+    index: number;
+    findedCity: string;
+    notes: NotesType[];
+    city: string | undefined;
+    removeNote: (title: string | undefined) => void;
+};
