+++ src/widgets/notes/components/accordionWithButtonForm/typings/index.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent } from "react";
-import { NotesType } from '../../../types';
+import { NotesType } from '../../../typings';
 
 export type PropsType = {
     title: string;
