+++ src/widgets/notes/components/button/typings/index.ts
