+++ src/widgets/notes/components/formWithNotes/typings/index.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent } from 'react';
-import { NotesType } from '../../../types';
+import { NotesType } from '../../../typings';
 
 export type PropsType = {
     title: string;
