+++ src/widgets/notes/typings/index.ts
